backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
gitpython>=3.1.0
google-generativeai>=0.3.0
pylint>=3.0.0
flake8>=6.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
requests>=2.31.0
